requirements.txt
Appium-Python-Client==0.50
asn1crypto==0.24.0
backports.shutil-get-terminal-size==1.0.0
cffi==1.14.0
chardet==3.0.4
cryptography==2.1.4
decorator==4.1.2
enum34==1.1.6
idna==2.9
ipaddress==1.0.17
ipython==5.5.0
ipython-genutils==0.2.0
jeepney==0.4.2
keyring==10.6.0
keyrings.alt==3.0
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.2.0
pyxdg==0.25
SecretStorage==2.3.1
selenium==3.141.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.25.8
wcwidth==0.1.7
